Added ADC interface and started new precharge sequence
