+++ Core/Test/tests.h
@@ -5,7 +5,7 @@
 #include <stdbool.h>
 #include <math.h>
 
-#include "io_control.h"
+#include "hv_control.h"
