+++ Core/User/hvc_config.h
@@ -16,20 +16,6 @@
 #define AIR_M_PIN GPIO_PIN_1
 #define PC_PIN GPIO_PIN_0
 
-#define ADC_BUFFER_LEN 8
-
-typedef enum {
-    ADC_VBATT,
-    ADC_VTS,
-    ADC_HALL_CH1,
-    ADC_HALL_CH2,
-    ADC_IMON,
-    ADC_SDC_IN,
-    ADC_SDC_LATCH,
-    ADC_MPU_TEMP,
-    ADC_NUM_SENSORS  // This is optional, but useful to denote the number of sensors
-} SensorType_t;
-
 
 // PRECHARGE SETTINGS
